chatuicomparison-UICode/requirements.txt
openai
python-dotenv
gradio